+++ R/f7Link.R
@@ -5,6 +5,9 @@
 #' @param label Optional link text.
 #' @param href Link source, url.
 #' @param icon Link icon, if any. Must pass \link{f7Icon}.
+#' @param routable Whether to make the link handled by
+#' the framework 7 router. Default to FALSE which opens
+#' a new page in a new tab.
 #'
 #' @examples
 #' if (interactive()) {
@@ -27,12 +30,12 @@
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #'
 #' @export
-f7Link <- function(label = NULL, href, icon = NULL) {
-  linkCl <- "link external"
+f7Link <- function(label = NULL, href, icon = NULL, routable = FALSE) {
+  linkCl <- if (routable) "link" else "link external"
 
   shiny::a(
     href = href,
-    target = "_blank",
+    target = if (!routable) "_blank",
     class = linkCl,
     if (!is.null(icon)) {
       shiny::tagList(