+++ inst/examples/gallery/tabs/tabBtns.R
@@ -33,7 +33,6 @@ tabBtns <- f7Tab(
 
   # segments
   f7Segment(
-    container = "row",
     f7Button(color = "blue", label = "My button", fill = FALSE),
     f7Button(color = "green", label = "My button", href = "https://www.google.com", fill = FALSE),
     f7Button(color = "yellow", label = "My button", fill = FALSE)
@@ -41,7 +40,6 @@ tabBtns <- f7Tab(
   f7BlockTitle(title = "Filled Buttons in a segment/rounded container"),
   f7Segment(
     rounded = TRUE,
-    container = "segment",
     f7Button(color = "black", label = "Action Button", inputId = "button2"),
     f7Button(color = "green", label = "My button", href = "https://www.google.com"),
     f7Button(color = "yellow", label = "My button")
@@ -49,21 +47,18 @@ tabBtns <- f7Tab(
   f7BlockTitle(title = "Outline Buttons in a segment/shadow container"),
   f7Segment(
     shadow = TRUE,
-    container = "segment",
     f7Button(label = "My button", outline = TRUE, fill = FALSE),
     f7Button(label = "My button", outline = TRUE, fill = FALSE),
     f7Button(label = "My button", outline = TRUE, fill = FALSE)
   ),
   f7BlockTitle(title = "Rounded Buttons in a segment container"),
   f7Segment(
-    container = "segment",
     f7Button(color = "blue", label = "My button", rounded = TRUE),
     f7Button(color = "green", label = "My button", rounded = TRUE),
     f7Button(color = "yellow", label = "My button", rounded = TRUE)
   ),
   f7BlockTitle(title = "Buttons of different size in a row container"),
   f7Segment(
-    container = "row",
     f7Button(color = "pink", label = "My button", shadow = TRUE),
     f7Button(color = "purple", label = "My button", size = "large", shadow = TRUE),
     f7Button(color = "orange", label = "My button", size = "small", shadow = TRUE)